pkg/config/configdoc: move DefaultsOnly
