+++ pkg/config/configtest/defaults.go
@@ -1,99 +1,15 @@
 package configtest
 
 import (
-	"bufio"
-	"fmt"
 	"io"
-	"reflect"
-	"strings"
+
+	"github.com/smartcontractkit/chainlink-common/pkg/config/configdoc"
 )
 
 // DocDefaultsOnly reads only the default values from a docs TOML file and decodes in to cfg.
 // Fields without defaults will set to zero values.
 // Arrays of tables are ignored.
+// Deprecated: use configdoc.DefaultsOnly
 func DocDefaultsOnly(r io.Reader, cfg any, decode func(io.Reader, any) error) error {
-	pr, pw := io.Pipe()
-	defer pr.Close()
-	go writeDefaults(r, pw)
-	if err := decode(pr, cfg); err != nil {
-		return fmt.Errorf("failed to decode default core configuration: %w", err)
-	}
-	// replace niled examples with zero values.
-	nilToZero(reflect.ValueOf(cfg))
-	return nil
-}
-
-// writeDefaults writes default lines from defaultsTOML to w.
-func writeDefaults(r io.Reader, w *io.PipeWriter) {
-	defer w.Close()
-	s := bufio.NewScanner(r)
-	var skipUntil func(line string) bool
-	for s.Scan() {
-		t := s.Text()
-		if skipUntil != nil {
-			if skipUntil(t) {
-				skipUntil = nil
-			}
-			continue
-		}
-		// Skip comments and examples (which become zero values)
-		if strings.HasPrefix(t, "#") || strings.HasSuffix(t, "# Example") {
-			continue
-		}
-		// Skip arrays of tables
-		if strings.HasPrefix(t, "[[") {
-			// skip fields until next empty line
-			skipUntil = func(line string) bool { return strings.TrimSpace(line) == "" }
-			continue
-		}
-		if _, err := io.WriteString(w, t); err != nil {
-			w.CloseWithError(err)
-		}
-		if _, err := w.Write([]byte{'\n'}); err != nil {
-			w.CloseWithError(err)
-		}
-	}
-	if err := s.Err(); err != nil {
-		w.CloseWithError(fmt.Errorf("failed to scan core defaults: %w", err))
-	}
-}
-
-func nilToZero(val reflect.Value) {
-	if val.Kind() == reflect.Ptr {
-		if val.IsNil() {
-			t := val.Type().Elem()
-			val.Set(reflect.New(t))
-		}
-		if val.Type().Implements(textUnmarshalerType) {
-			return // don't descend inside - leave whole zero value
-		}
-		val = val.Elem()
-	}
-	switch val.Kind() {
-	case reflect.Struct:
-		if val.Type().Implements(textUnmarshalerType) {
-			return // skip values unmarshaled from strings
-		}
-		for i := 0; i < val.NumField(); i++ {
-			f := val.Field(i)
-			nilToZero(f)
-		}
-		return
-	case reflect.Map:
-		if !val.IsNil() {
-			for _, k := range val.MapKeys() {
-				nilToZero(val.MapIndex(k))
-			}
-		}
-		return
-	case reflect.Slice:
-		if !val.IsNil() {
-			for i := 0; i < val.Len(); i++ {
-				nilToZero(val.Index(i))
-			}
-		}
-		return
-	default:
-		return
-	}
+	return configdoc.DefaultsOnly(r, cfg, decode)
 }
